Add more tests
Now the scripts check the behaviour of an implementation if the given hash is too small (done if an adler32 hash of the email address), if size option has a negative value, if the default option has an unimplemented value and if this unimplemented value is forced (with forcedefault=y).

While here the _md5 function was moved to regress.sh.